requirements.txt
Babel==1.3
Django==1.6.5
Pillow==2.5.3
South==0.8.4
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.18
boto==2.27.0
celery==3.1.15
cffi==0.8.2
cryptography==0.4
dj-database-url==0.2.2
dj-static==0.0.5
django-appconf==0.6
django-braces==1.4.0
django-celery==3.1.10
django-extra-fields==0.1
django-filter==0.7
django-mandrill==0.1.0
django-oauth-toolkit==0.7.2
django-storages==1.1.8
django-toolbelt==0.0.1
djangorestframework==2.3.13
djangorestframework-bulk==0.1.4
docopt==0.4.0
easy-thumbnails==2.1
elasticsearch==0.4.5
elasticutils==0.9.1
gunicorn==18.0
httplib2==0.9
kombu==3.0.21
mandrill==1.0.56
numpy==1.8.1
oauth2client==1.2
oauthlib==0.6.3
psycopg2==2.5.3
pyOpenSSL==0.14
pycparser==2.10
pycrypto==2.6.1
pystache==0.5.3
python-dateutil==2.2
python-magic==0.4.6
pytz==2014.2
redis==2.9.1
requests==2.3.0
rq==0.3.13
six==1.8.0
smtpapi==0.1.2
static==1.0.2
times==0.6.2
unittest2==0.5.1
urllib3==1.9
wsgiref==0.1.2
